+++ NEWS.md
@@ -1,3 +1,8 @@
+# RcppCWB 0.6.3.9001
+
+* `cwb_huffcode()` and `cwb_compress_rdx()` did not delete redundant files on 
+Windows. Fixed by temporarily unloading the corpus #89.
+
 # RcppCWB 0.6.3
 
 * `cl_struc_values()` does not duplicate registry directories any more #77.